streamlit_app.py

import streamlit as st
from PIL import Image

st.title("Facial Mapping AI - MVP")

uploaded_file = st.file_uploader("Upload a front-facing image", type=["jpg", "jpeg", "png"])

if uploaded_file is not None:
    image = Image.open(uploaded_file)
    st.image(image, caption="Uploaded Image", use_column_width=True)

    # Simulate facial mapping overlay
    st.markdown("### Facial landmarks and proportion lines would be displayed here (placeholder)")
    st.image(image, caption="Processed Image (Simulated Overlays)", use_column_width=True)
    st.success("Image processed and ready for consult export.")
